Appearance block in AAM, React code refactoring
