+++ cvat-ui/src/components/annotation-page/standard-workspace/standard-workspace.tsx
@@ -8,9 +8,9 @@ import Layout from 'antd/lib/layout';
 
 import CanvasWrapperContainer from 'containers/annotation-page/standard-workspace/canvas-wrapper';
 import ControlsSideBarContainer from 'containers/annotation-page/standard-workspace/controls-side-bar/controls-side-bar';
-import ObjectSideBarContainer from 'containers/annotation-page/standard-workspace/objects-side-bar/objects-side-bar';
 import PropagateConfirmContainer from 'containers/annotation-page/standard-workspace/propagate-confirm';
 import CanvasContextMenuContainer from 'containers/annotation-page/standard-workspace/canvas-context-menu';
+import ObjectSideBarComponent from 'components/annotation-page/standard-workspace/objects-side-bar/objects-side-bar';
 import CanvasPointContextMenuComponent from 'components/annotation-page/standard-workspace/canvas-point-context-menu';
 
 export default function StandardWorkspaceComponent(): JSX.Element {
@@ -18,7 +18,7 @@ export default function StandardWorkspaceComponent(): JSX.Element {
         <Layout hasSider className='cvat-standard-workspace'>
             <ControlsSideBarContainer />
             <CanvasWrapperContainer />
-            <ObjectSideBarContainer />
+            <ObjectSideBarComponent />
             <PropagateConfirmContainer />
             <CanvasContextMenuContainer />
             <CanvasPointContextMenuComponent />
